[DMS-758] Use int32 instead of int16 for the TotalCount result (#627)

+++ src/dms/backend/EdFi.DataManagementService.Backend.Postgresql/Operation/SqlAction.cs
@@ -438,7 +438,7 @@ public partial class SqlAction() : ISqlAction
         }
 
         await reader.ReadAsync();
-        return reader.GetInt16(reader.GetOrdinal("Total"));
+        return reader.GetInt32(reader.GetOrdinal("Total"));
     }
 
     /// <summary>
